added in funcitonalily for fitting with seperate scans
I added in both a fitter and a x0 guesser that takes a
a seperature input for gain and fine scans. This
should be more robust becasue the guesser doesn't
have to figure out what is a gain scan and what is
a fine scan. Also now the fitter doesn't require to
premeptivily stack the fine and gain scans together.