added in ploting coherence between detectors for PCA
I added in two plots in PCA_diag.pdf one is a plot of two psds and
the cross psd for the two most correlated detectors. The other is
Then a plot of the coherence between the two detectors as a
function of frequency. I tested this on a noise set where the
coherence was low everywhere except for the pulse tube frequency
and a 60Hz frequency. Then I tested it on a beam map time stream
and the coherence was high for the 1/f signal and for the 5Hz
chop signal