fit_fine_gain now flags data near other resonators
I added in not using the data that near other resonators
when for fitting. for the fine scan data it only uses the
data halfway inbetween two resonators if they are very close
to each other. For the gain data it does not use any
frequency points that are within one fine scan span of a
resonator