faster root finding and brute force linear mag fit
I added in faster root finding for all of the nonlinear
fits. This was the biggest slow down area. I did this
by using the analitic method for finding roots to the
the cubic equation and implamenting with numba which
executes without slow python. In addition I added a
vectorized brute for fitter for fitting in linear
magnitude space. It can produce a plot that shows
the covariance using delta chi squared