Added the models for disordered superconductivity
Added the 'disordered_model.py' as an implementation of the
Abrikosov-Gorkov model similar to the approaches in Pieter Coumou's
thesis and Driessen et al. 2012
'dynes_model.py' takes a similar approach, but for a Dynes type
broadening density of states model mostly following the Nam approach in
Zemlicka et al. 2015