quality of life improvements for resonance fitting
changes to resonance_fitting
I cleaned up the print statements for guessing and added a print
statement for the fitted results.
Into the fit dictionaries I added in keys that have names of the
variables so that the results are more easily human readable and
accesable.
when specifying tau in fit_nonlinear_iq I know broadcast the fit
result and covariance matrix into arrays with the same shape as if
you had fit tau
I fixed a bug in fit_nonlinear_mag where x0_result and fit_result
where in magnitude squared instead of magnitude. This required you
to have to plot the data as 20*np.log10(np.abs(z)) but then plot
the results as 10*np.log10(np.abs(z)). Unfortunaly this will ruin
the plotting for old code that did just plot things like this.
I added in the fitting in magnitude space to the res_fit demo