allow interactive iq sweep plotter to have multiple traces
updated the interative iq sweep plotter to allow you to have
n traces/sweeps. That way it can be used to display fits as in the
example or it could be used to display a sweep over some parameter
like blackbody or array temperature. Also added legned for the
combined data that displays the value and the ability to double
click to just to a particual index resonator