fix autoscaling if there are values that are infinite
If you resonator is very strongly Qc limited calculation of Qi
will not be very good and you can end up with very large Qi
values and sometimes they will be infinite. That was breaking
the autoscaling on the interactive plot. So this is just a fix
for that situation