fixed guessing bug for really deep resonators
amp_guess uses a polynomial fit to the depth of a resonator
versus the amp value. This fit was made for 0 to 30dB deep
resonators. For really deep  resonators it was returning negative
values. Negative values are infeasible solutions.
This change fixes amp_guess to 0.99 for depths greater
than 30dB